refactor(blocks): extract sample primitives from preview blocks
Pull the per-token renderers out of MotionPreview, ShadowPreview,
BorderPreview, and DimensionScale into standalone components so MDX
authors can embed a single sample:

- MotionSample (path, speed?, runKey?)
- ShadowSample (path)
- BorderSample (path)
- DimensionBar (path, kind?)

The parent blocks stay unchanged in DOM output and props — they're now
thin iterators that filter, sort, and map over the extracted sample.
Adds a MotionSample story and refreshes the Blocks table + reference
docs with the new entries.

Closes #175
